requirements.txt
fastapi
uvicorn
python-jose
python-multipart
pillow
pydantic
email-validator
google-auth
